+++ apps/web-guy-portal/src/components/Page.tsx
@@ -1,4 +1,13 @@
-import { Card, CardProps, GridSection, GridSectionProps, Header, ModulesComponents } from 'library-sb';
+import {
+  Card,
+  CardProps,
+  RichText,
+  Header,
+  ModulesComponents,
+  RichTextProps,
+  ExperienceSectionProps,
+  ExperienceSection,
+} from 'library-sb';
 import React, { ReactNode, useEffect, useState } from 'react';
 import { getPage } from '../contentful/get-page';
 import { ContentfulPage } from '../types/page-types';
@@ -6,7 +15,8 @@ import { ContentfulPage } from '../types/page-types';
 const renderModuleComponent = (modulesList: string[], args: CardProps[]): ReactNode[] => {
   const modulesComponents: ModulesComponents = {
     mlCard: Card,
-    orGridSection: GridSection,
+    richText: RichText,
+    mlExperienceSection: ExperienceSection,
   };
 
   return modulesList.map((module: string, index: number) => (
@@ -17,14 +27,14 @@ const renderModuleComponent = (modulesList: string[], args: CardProps[]): ReactN
 const Page = () => {
   const [page, setPage] = useState<ContentfulPage | undefined>(undefined);
   const [modules, setModules] = useState<string[]>([]);
-  const [moduleArgs, setModuleArgs] = useState<CardProps[]>([]);
+  const [moduleArgs, setModuleArgs] = useState<(CardProps | RichTextProps | ExperienceSectionProps)[]>([]);
 
-  const mapContentTypesModules = (modulesArray: (CardProps | GridSectionProps)[]) => {
+  const mapContentTypesModules = (modulesArray: (CardProps | RichTextProps | ExperienceSectionProps)[]) => {
     for (let index = 0; index < modulesArray.length; index++) {
       const { contentTypeId, ...args } = modulesArray[index];
 
       setModules((prevModules: string[]) => [...prevModules, contentTypeId || '']);
-      setModuleArgs((prevArgs: (CardProps | GridSectionProps)[]) => [...prevArgs, args]);
+      setModuleArgs((prevArgs: (CardProps | RichTextProps | ExperienceSectionProps)[]) => [...prevArgs, args]);
     }
   };
 
@@ -41,9 +51,9 @@ const Page = () => {
   }, []);
 
   return (
-    <div>
+    <div className="grid grid-cols-4 mx-10">
       <Header categories={page?.header?.categories} hero={page?.header?.hero} languajes={page?.header?.languajes} />
-      <main className="flex flex-col min-h-screen space-y-1/2 mt-1/10">
+      <main className="col-span-full min-h-screen space-y-1/3 mx-20">
         {modules.length ? renderModuleComponent(modules, moduleArgs) : <></>}
       </main>
     </div>