+++ apps/web-guy-portal/src/types/page-types.ts
@@ -1,9 +1,9 @@
-import { CardProps, GridSectionProps, HeaderProps, RichTextProps } from 'library-sb';
+import { CardProps, HeaderProps, RichTextProps, ExperienceSectionProps } from 'library-sb';
 
 export interface ContentfulPage {
   title?: string;
   baseUrl?: string;
   header?: HeaderProps;
-  modules?: (CardProps | GridSectionProps)[] | undefined;
+  modules?: (CardProps | RichTextProps | ExperienceSectionProps)[] | undefined;
   footer?: unknown;
 }