+++ packages/library-sb/src/main.tsx
@@ -1,10 +1,11 @@
 export * from './stories/at-button/Button';
 export * from './stories/at-image/Image';
+export * from './stories/at-richtext/RichText';
 
 export * from './stories/ml-card/Card';
 
+export * from './stories/or-experience-section/Experience-section';
 export * from './stories/or-header/Header';
-export * from './stories/or-grid-section/Grid-section';
 
 export * from './types/components-type-props';
 export * from './types/contentful-types';