+++ apps/web-guy-portal/CHANGELOG.md
@@ -1,5 +1,16 @@
 # web-guy-portal
 
+## 3.0.0
+
+### Major Changes
+
+- Add importan main section (experience section)
+
+### Patch Changes
+
+- Updated dependencies
+  - library-sb@3.0.0
+
 ## 2.0.0
 
 ### Major Changes