+++ packages/library-sb/CHANGELOG.md
@@ -1,5 +1,11 @@
 # library-sb
 
+## 3.0.0
+
+### Major Changes
+
+- Add importan main section (experience section)
+
 ## 2.0.0
 
 ### Major Changes